Create check registry with initial noop check, and vendor dependencies
Factor the KubeObjects type out of package main and into the top-level
clusterlint package. For now still populate it in main, though we'll
likely want to pull this out later.

Add a check type that checks will implement, and implement an initial
noop check that satisfies it. Create a registry for checks.

Vendor dependencies so that we can do a consistent build. For now we're
depending on the master version of client-go since they don't have a
release that supports modules yet, but we should move to a release once
they do.