Add a DOKS check for custom node labels and taints
In DOKS labels and taints applied to nodes will be lost when the cluster
is upgraded or a node is otherwise replaced. This can cause problems for
workloads if labels or taints are used for scheduling.

Add a warning if any node in a cluster has custom labels or taints.