Error on images from docker.pkg.github.com
Since Kubernetes 1.20 containerd is used instead of Docker as the container runtime.
Containerd is due to protocol version differences/support/mismatch unabled to
pull images hosted at docker.pkg.github.com. The new check in this commit will
error when it finds an image from that registry, and suggests to use ghcr.io,
which is also hosted and operated by GitHub, and the successor of
docker.pkg.github.com.

Refs:
* https://github.com/containerd/containerd/issues/3291#issuecomment-683700425
* https://docs.github.com/en/packages/guides/migrating-to-github-container-registry-for-docker-images#domain-changes