Ignore "not found" errors when listing resources
If a resource version doesn't exist in a cluster, there can't be any objects of
that kind for us to check, so ignore the error when trying to list them.

While we're here, enhance the errors we get when listing objects so that we can
tell which object kind failed.